refactored code into separate modules #18
